Move getMobileOSName into common utils file
