+++ src/components/Capture/Document.js
@@ -11,10 +11,9 @@ import withCameraDetection from './withCameraDetection'
 import withCrossDeviceWhenNoCamera from './withCrossDeviceWhenNoCamera'
 import withHybridDetection from './withHybridDetection'
 import { getDocumentTypeGroup } from '../DocumentSelector/documentTypes'
-import { isDesktop, addDeviceRelatedProperties } from '~utils'
+import { isDesktop, addDeviceRelatedProperties, getMobileOSName } from '~utils'
 import { compose } from '~utils/func'
 import { randomId, upperCase } from '~utils/string'
-import { getMobileOSName } from '~utils/detectMobileOS'
 import { localised } from '../../locales'
 import style from './style.css'
 import FallbackButton from '../Button/FallbackButton'
