+++ src/components/Capture/Face.js
@@ -10,10 +10,9 @@ import withCrossDeviceWhenNoCamera from './withCrossDeviceWhenNoCamera'
 import GenericError from '../GenericError'
 import FallbackButton from '../Button/FallbackButton'
 import CustomFileInput from '../CustomFileInput'
-import { isDesktop, addDeviceRelatedProperties } from '~utils'
+import { isDesktop, addDeviceRelatedProperties, getMobileOSName } from '~utils'
 import { compose } from '~utils/func'
 import { randomId, upperCase } from '~utils/string'
-import { getMobileOSName } from '~utils/detectMobileOS'
 import { getInactiveError } from '~utils/inactiveError.js'
 import { localised } from '../../locales'
 import style from './style.css'
