+++ src/components/utils/index.js
@@ -24,7 +24,20 @@ export const preventDefaultOnClick = callback => event => {
 
 // iPad 13 platform is 'MacIntel'
 // https://stackoverflow.com/questions/58019463/how-to-detect-device-name-in-safari-on-ios-13-while-it-doesnt-show-the-correct
-let isIOS = (/iPad|iPhone|iPod/.test(navigator.platform || '') || (navigator.platform === 'MacIntel' && navigator.maxTouchPoints > 1)) && !window.MSStream
+const isIOS = (/iPad|iPhone|iPod/.test(navigator.platform || '') || (navigator.platform === 'MacIntel' && navigator.maxTouchPoints > 1)) && !window.MSStream
+
+// WARN: use of this util and navigator.userAgent is highly discouraged unless absolutely necessary and for simple use cases
+export const getMobileOSName = () => {
+  console.warn("getMobileOSName - use of navigator.userAgent is highly discouraged unless absolutely necessary and only for simple use cases")
+  const userAgent = navigator.userAgent
+  if (/android/i.test(userAgent)) {
+    return "Android"
+  }
+  if (/iPad|iPhone|iPod/i.test(userAgent)) {
+    return "iOS"
+  }
+  console.error("Unable to determine mobile OS")
+}
 
 // Copied from https://github.com/muaz-khan/DetectRTC/blob/master/DetectRTC.js
 export const isDesktop = !(/Android|webOS|BB10|BlackBerry|IEMobile|Opera Mini|Mobile|mobile/i.test(navigator.userAgent || '')) && !isIOS
